Allocation and Sharing UI updates - Commented out code for Copy allocation image - Updated UI for legend
Shrijan, I got started with the pie chart but it is not done yet. I think we need to add code to change the size depending on the viewport size (similar to what you did in the home page for blogs) When the viewport is small the chart size will be smaller than 400 x 400.